Don't include services in uber jar
The Web Gateway currently builds into an uber jar that
can be run via java -jar.

This patch changes dependencies to scope 'provided'
so that they are not included in the jar. The services in
question are .war files which are loaded during run-time.

Reviewed-by: stooke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-April/022618.html