Update jvm-memory service with implementations for PUT, POST and DELETE
This patch implements PUT, POST and DELETE (GET is already
implemented) for the jvm-memory service. The service is now fully
functional and respects the API defined in jvm-memory-swagger.yaml
This is now usable by the Agent to input data, as well
as clients to access data as before.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-May/022942.html