Fall back to system property if THERMOSTAT_GATEWAY_HOME not set in environment
If the environment variable THERMOSTAT_GATEWAY_HOME is not set, this patch checks to see if it's set as a Java system property.
This is useful  as it allows another level of configuration - the command line should always take precedence over environment.
This behaviour is consistent with the old Thermostat.

Reviewed-by: sgehwolf
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-June/023498.html